ai-service/requirements.txt
# Money OS — AI Service Dependencies
# Install with: pip install -r requirements.txt

# FastAPI web framework
fastapi==0.110.0

# ASGI server (run with: uvicorn main:app --reload)
uvicorn[standard]==0.27.1

# Data validation & serialization
pydantic==2.6.3

# HTTP client (for calling Spring Boot if needed)
httpx==0.27.0

# CORS middleware (already included in fastapi, but explicit)
starlette==0.36.3

# Data science / math
numpy==1.26.4
pandas==2.2.1

# Simple ML (no complex dependencies)
scikit-learn==1.4.1.post1

# Environment variables
python-dotenv==1.0.1

# Date utilities
python-dateutil==2.9.0
